Added effect-parsers
- Effect parsers: parsers for base effects, their container, and a country modifier builder are implemented. "KnownValues" text file contains the known effects, units etc.
- Slight DB schema changes: long for storage of pearls and coral. Effect has a name. Added some comments.
- Removed the migrations folder, added it to gitignore (DB is local DB for everyone, migrations should be handled individually)

+++ backend/StrategyGame.Model/Entities/Country.cs
@@ -11,12 +11,12 @@ namespace StrategyGame.Model.Entities
         /// <summary>
         /// Gets or sets the current amount of pearls the country has.
         /// </summary>
-        public int Pearls { get; set; }
+        public long Pearls { get; set; }
 
         /// <summary>
         /// Gets or sets the current amount of corals the country has.
         /// </summary>
-        public int Corals { get; set; }
+        public long Corals { get; set; }
 
         /// <summary>
         /// Gets the user to whom the country belongs.
@@ -66,7 +66,7 @@ namespace StrategyGame.Model.Entities
         /// </summary>
         /// <param name="parentUser">The user to whom the country belongs.</param>
         /// <param name="initialPearls">The initial amount of pearls the country has.</param>
-        /// <param name="initialCorals">The initial amount of corals the country has.></param>
+        /// <param name="initialCorals">The initial amount of corals the country has.</param>
         /// <exception cref="ArgumentNullException">Thrown if an argument was null.</exception>
         public Country(User parentUser, int initialPearls, int initialCorals)
         {
